[ Matti-Dev ] FINAL feature added; Charts -> few adjustments
In the dashboard and home page are now all price charts displayed as intented. I have changed the stylings for better readability of the chart information. At first, the plots where encoded in the wrong backend mode, I changed it to 'AGG' along the matplotlib docs to fix it.
