+++ amazon_price_tracker_app/data/urls.txt
@@ -1,2 +1,3 @@
-https://fakestoreapi.com/products/1
-https://fakestoreapi.com/products/3
+https://fakestoreapi.com/products/2
+https://fakestoreapi.com/products/4
+https://fakestoreapi.com/products/16
