+++ amazon_price_tracker_app/templates/dashboard.html
@@ -27,10 +27,10 @@
                         </div>
                   </div>
                   {% comment %} price development chart {% endcomment %}
-                  <div class="h-full w-full row-span-3 col-span-1 p-3">
+                  <div class="h-full w-full row-span-3 col-span-1 flex items-center justify-center pt-3">
                         <img src="data:image/png;base64,{{ prod_plot_string }}"
                              alt="{{ product.name }}"
-                             class="w-full h-full rounded-xl"
+                             class="h-full rounded-xl aspect-video"
                              height=""
                              width="">
                   </div>
